requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pydantic
jinja2
python-multipart
